refactor/categories-section
