+++ src/app/categories/page.js
@@ -1,7 +1,8 @@
 
 import Footer from "@/components/footer/footer";
-import CategorieCard from "./categorieCard";
-import Filtercategorie, { Pagination, Paginationbar } from "./filter";
+import CategorieCard from "@/components/categories/categorie-card";
+import { Paginationbar } from "@/components/pagination/pagination-bar";
+import CategorieFilter from "@/components/categories/categorie-filter";
 
 export default function Categories() {
   const Categories = [
@@ -63,7 +64,7 @@ export default function Categories() {
               Showing 24 categories
             </p>
             <div>
-              <Filtercategorie />
+              <CategorieFilter />
             </div>
           </div>
           <div className=" grid grid-cols-2 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4  sm:gap-9 w-full m-0 mt-9 h-full ">
