+++ src/components/about/about-sections-images.js
